moved joypad again, changed variable name for playing sound in application
